requirements.txt
cryptography==35.0.0
django==3.2.8
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
drf-spectacular==0.20.2
freezegun==1.1.0
gunicorn==20.1.0
phonenumbers==8.12.35
psycopg2-binary==2.9.1
pyjwt[crypto]==2.3.0
requests==2.27.1
twilio==7.3.0
whitenoise==5.3.0
python-dateutil==2.8.2
